Rearange of directories and components
