+++ src/App.js
@@ -4,7 +4,7 @@ import Materias from './components/Materias'
 import ContactForm from './components/Contact'
 import Home from './components/Home'
 import Login from './components/Login'
-import NotFound from './shared/NotFound'
+import NotFound from './components/NotFound'
 import Footer from './components/Footer'
 import Novedades from './components/Novedades'
 import { AuthContext } from './context/AuthContext'
@@ -72,13 +72,7 @@ function App() {
           <Route path='/home' element={<Home />} />
           <Route path='/login' element={<Login onSubmit={onSubmitLogin} />} />
           <Route path='/Novedades' element={<Novedades />} />
-
-          <Route
-            path='/Materias'
-            element={
-                <Materias />
-            }
-          />
+          <Route path='/Materias'element={<Materias />}/>
           <Route path='/Contacto' element={<ContactForm />} />
           <Route path='*' element={<NotFound />} />
         </Routes>
