+++ src/components/Layout.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Navbar from '../shared/Navbar'
+import Navbar from './Navbar'
 
 const Layout = ({ children }) => {
   return (
