--- src/shared/Navbar.js
@@ -1,89 +1,0 @@
-import React from 'react'
-import { NavLink } from 'react-router-dom'
-import { useState } from 'react'
-
-const Navbar = () => {
-  let activeStyle = {
-    textDecoration: 'underline',
-    color: 'blue',
-  }
-
-  const [navbar, setNavbar] = useState(false)
-  return (
-    <header>
-      <nav className='w-full bg-gray-800 shadow'>
-        <div className='justify-between px-4 mx-auto lg:max-w-7xl md:items-center md:flex md:px-8 '>
-          <div>
-            <div className='flex items-center justify-between py-3 md:py-5 md:block'>
-              <a href='/Home'>
-                <h2 className='text-2xl text-white font-bold py-3'>SITEC</h2>
-              </a>
-              <div className='md:hidden'>
-                <button
-                  className='p-2 text-gray-700 rounded-md outline-none focus:border-gray-400 focus:border'
-                  onClick={() => setNavbar(!navbar)}
-                >
-                  {navbar ? (
-                    <svg
-                      xmlns='http://www.w3.org/2000/svg'
-                      className='w-6 h-6 text-white'
-                      viewBox='0 0 20 20'
-                      fill='currentColor'
-                    >
-                      <path
-                        fillRule='evenodd'
-                        d='M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z'
-                        clipRule='evenodd'
-                      />
-                    </svg>
-                  ) : (
-                    <svg
-                      xmlns='http://www.w3.org/2000/svg'
-                      className='w-6 h-6 text-white'
-                      fill='none'
-                      viewBox='0 0 24 24'
-                      stroke='currentColor'
-                      strokeWidth={2}
-                    >
-                      <path
-                        strokeLinecap='round'
-                        strokeLinejoin='round'
-                        d='M4 6h16M4 12h16M4 18h16'
-                      />
-                    </svg>
-                  )}
-                </button>
-              </div>
-            </div>
-          </div>
-          <div>
-            <div className='text-slate-500 uppercase space-x-6 font-semibold'>
-              <NavLink to='/home' style={({ isActive }) => (isActive ? activeStyle : undefined)}>
-                Home
-              </NavLink>
-
-              <NavLink
-                to='/Novedades'
-                style={({ isActive }) => (isActive ? activeStyle : undefined)}
-              >
-                Novedades
-              </NavLink>
-
-              <NavLink
-                to='/Contacto'
-                style={({ isActive }) => (isActive ? activeStyle : undefined)}
-              >
-                Contacto
-              </NavLink>
-              <NavLink to='/Login' style={({ isActive }) => (isActive ? activeStyle : undefined)}>
-                Login
-              </NavLink>
-            </div>
-          </div>
-        </div>
-      </nav>
-    </header>
-  )
-}
-
-export default Navbar
